+++ libs/sdk-core/CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Improvements
 - Manage properly 529 status (service overloaded) on chat endpoint
 - Add `getEntityFromFilter` and `getFilterFromEntity` functions
+- Add `autofilters` property to `FindResults` model
 
 # 1.3.4 (2023-06-13)